+++ NEWS.md
@@ -12,6 +12,7 @@ New Features(Analysis):
 - Properly detect `PhanUnusedVariable` in try-catch where catch always rethrows. (#4567)
 - Make read-only/write-only property detection more accurate for assignment operations (e.g. `+=`, `??=`) and increment/decrement operations. (#4570)
 - Improve estimates of array sizes when analyzing calls that unpack values, based on the inferred real type set. (#4577)
+- Emit `PhanParamTooFewUnpack` and `PhanParamTooFewInternalUnpack` to indicate when argument unpacking may provide too few arguments to the called function. (#4577)
 - Support `@phan-type AliasName=UnionType` annotation in inline strings or element comments (#4562)
 
   These aliases will apply to remaining statements in the current